part 2

+++ interface/super/edit_globals.php
@@ -126,9 +126,38 @@ if ($_POST['form_save'] && $_GET['mode'] == "user") {
       foreach ($grparr as $fldid => $fldarr) {
         if (in_array($fldid, $USER_SPECIFIC_GLOBALS)) {
           list($fldname, $fldtype, $flddef, $flddesc, $fldlist) = $fldarr;
+          //check and validate input from client side with globals.
+          if(is_array($grparr[$fldid][1])) {
+          $search = $grparr[$fldid][1];
+          $boolean = array_key_exists(trim(strip_escape_custom($_POST["form_$i"])),$search);
+          }
+          elseif ($grparr[$fldid][1] == bool) {
+          $_POST["form_$i"] == 0;
+          $boolean = true;
+          }
+          elseif ($fldid == "css_header") {
+            //styles array created since the globals does not have styles array.
+            $styles_array = array("style_prism.css", "style_light.css", "style_purple.css", "style_tan.css", "style_tan_no_icons.css");
+            if (in_array($_POST["form_$i"], $styles_array)) {
+              $boolean = true;
+            }
+            else {
+              $boolean = false;
+            }
+          }
+          elseif ($fldid == "primary_color" || $fldid == "primary_font_color" || $fldid == "secondary_color" || $fldid == "secondary_font_color" ) {
+            if (strlen($_POST['form_$i']) == 7 && substr($_POST['form_$i'], 0,1) == "#") {
+            $boolean = true;
+            }
+            else {
+              $boolean = "false";
+            }
+          }
+          if ($boolean) { 
           $label = "global:".$fldid;
           $fldvalue = trim(strip_escape_custom($_POST["form_$i"]));
           setUserSetting($label,$fldvalue,$_SESSION['authId'],FALSE);
+          }
           if ( $_POST["toggle_$i"] == "YES" ) {
             removeUserSetting($label);
           }
@@ -225,9 +254,77 @@ if ($_POST['form_save'] && $_GET['mode'] != "user") {
                 sqlStatement("UPDATE libreehr_module_vars SET pn_value = ? WHERE pn_name = 'pcFirstDayOfWeek'", array($fldvalue));
                 break;
             }
+          //check and validate input from client side with globals.
+          if(is_array($grparr[$fldid][1])) {
+              $search = $grparr[$fldid][1];
+             $boolean = array_key_exists(trim(strip_escape_custom($_POST["form_$i"])),$search);
+
+          }
+          elseif ($grparr[$fldid][1] == "text") {
+            //for text fields
+              $_POST["form_$i"] = trim(strip_escape_custom($_POST["form_$i"]));
+              $boolean = true;
+          }
+          elseif ($grparr[$fldid][1] == bool) {
+          $_POST["form_$i"] == 0;
+          $boolean = true;
+          }
+          elseif ($fldid == "css_header") {
+            //styles array created since the globals does not have styles array.
+            $styles_array = array("style_prism.css", "style_light.css", "style_purple.css", "style_tan.css", "style_tan_no_icons.css");
+            if (in_array($_POST["form_$i"], $styles_array)) {
+              $boolean = true;
+            }
+            else {
+              $boolean = false;
+            }
+          }
+
+          elseif ($fldid == "primary_color" || $fldid == "primary_font_color" || $fldid == "secondary_color" || $fldid == "secondary_font_color" ) {
+            if (strlen($_POST['form_$i']) == 7 && substr($_POST['form_$i'], 0,1) == "#") {
+            $boolean = true;
+            }
+            else {
+              $boolean = "false";
+            }
+          }
+          elseif ($fldid == "theme_tabs_layout") {
+            //menu array created sine the globals dont have the array
+            $menu_array = array('tabs_style_compact.css', 'tabs_style_full.css');
+            if (in_array($_POST["form_$i"], $menu_array)) {
+              $boolean = true;
+            }
+            else {
+              $boolean = false;
+            }
+          }
+          elseif ($fldid == "language_default" OR $fldid="language_menu_other") {
+            $total_languages = sqlStatement('SELECT COUNT(*) FROM `lang_languages`');
+            if($_POST['form_$i'] <= $total_languages) {
+              $boolean = true;
+            }
+            else {
+              $boolean = false;
+            }
+          }
+          elseif ($fldid == "schedule_end" OR $fldid == "schedule_start") {
+            //we rely on face that time wont exceed 24 hrs
+            if ($_POST['form_$i'] < 24) {
+              $boolean = true;
+            }
+            else {
+              $boolean = false;
+            }
+          }
+
+          //boolean will determine whether the data is valid, else dont update.
+          if ($boolean) {
             // Replace old values
             sqlStatement( 'DELETE FROM `globals` WHERE gl_name = ?', array( $fldid ) );
+
             sqlStatement( 'INSERT INTO `globals` ( gl_name, gl_index, gl_value ) VALUES ( ?, ?, ? )', array( $fldid, 0, $fldvalue )  );
+          }
+
         } else {
           //error_log("No need to update $fldid");
         }
@@ -256,7 +353,7 @@ if ($_POST['form_save'] && $_GET['mode'] != "user") {
 <!-- supporting javascript code -->
 <?php
    // Including Bootstrap and Fancybox.
-   call_required_libraries(true,true,false,false);
+  call_required_libraries(array("jquery-min-3-1-1","bootstrap","fancybox"));
    include_js_library("jscolor-1-4-5/jscolor.js");
 ?>
 
@@ -283,18 +380,21 @@ input     { font-size:10pt; }
   <form method='post' name='theform' id='theform' action='edit_globals.php' onsubmit='return top.restoreSession()'>
 <?php } ?>
 
-<?php if ($_GET['mode'] == "user") { ?>
-  <p><b><?php echo xlt('Edit User Settings'); ?></b>
-<?php } else { ?>
-  <p><b><?php echo xlt('Edit Global Settings'); ?></b>
-<?php } ?>
-
+<div style="display:none">
+  <?php if ($_GET['mode'] == "user") { ?>
+    <p><b><?php echo xlt('Edit User Settings'); ?></b>
+  <?php } else { ?>
+    <p><b><?php echo xlt('Edit Global Settings'); ?></b>
+  <?php } ?>
+</div>
 <?php // mdsupport - Optional server based searching mechanism for large number of fields on this screen. ?>
-<span style='float: right;'>
-    <input name='srch_desc' size='20'
+<div style="float: right;">
+  <input type='submit' style="float: right;" name='form_search' value='<?php echo xla('Search'); ?>' />
+</div>
+<div style='float: right;'>
+    <input name='srch_desc' type="text" class="form-rounded form-control" size='20'
         value='<?php echo (!empty($_POST['srch_desc']) ? htmlspecialchars($_POST['srch_desc']) : '') ?>' />
-    <input type='submit' name='form_search' value='<?php echo xla('Search'); ?>' />
-</span>
+</div>
 
 <!--tabNav-->
 <ul class="tabNav">
@@ -309,7 +409,7 @@ foreach ($GLOBALS_METADATA as $grpname => $grparr) {
 ?>
 </ul> 
 
-<div class="tabContainer well">
+<div class="tabContainer well" style="height: 75%; overflow: auto;">
 <?php
 $i = 0;
 foreach ($GLOBALS_METADATA as $grpname => $grparr) {
@@ -364,7 +464,7 @@ foreach ($GLOBALS_METADATA as $grpname => $grparr) {
     echo " <tr $srch_cl title='" . attr($flddesc) . "'  id='".attr($fldid)."' value='".attr($fldvalue)."'><td><b>" . text($fldname) . "</b></td><td>\n";
 
     if (is_array($fldtype)) {
-      echo "  <select class='form-control' name='form_$i' id='form_$i'>\n";
+      echo "  <select class='form-control input-sm' name='form_$i' id='form_$i'>\n";
       foreach ($fldtype as $key => $value) {
         if ($_GET['mode'] == "user") {
           if ($globalValue == $key) $globalTitle = $value;
@@ -399,7 +499,7 @@ foreach ($GLOBALS_METADATA as $grpname => $grparr) {
       if ($_GET['mode'] == "user") {
         $globalTitle = $globalValue;
       }
-      echo "  <input type='text' class='form-control form-rounded' name='form_$i' id='form_$i' " .
+      echo "  <input type='text' class='form-control input-sm' name='form_$i' id='form_$i' " .
         "size='6' maxlength='15' value='" . attr($fldvalue) . "' />\n";
     }
 
@@ -407,7 +507,7 @@ foreach ($GLOBALS_METADATA as $grpname => $grparr) {
       if ($_GET['mode'] == "user") {
         $globalTitle = $globalValue;
       }
-      echo "  <input type='text' class='form-control form-rounded' name='form_$i' id='form_$i' " .
+      echo "  <input type='text' class='form-control input-sm' name='form_$i' id='form_$i' " .
         "size='50' maxlength='255' value='" . attr($fldvalue) . "' />\n";
     }
     else if ($fldtype == 'pwd') {
@@ -428,7 +528,7 @@ foreach ($GLOBALS_METADATA as $grpname => $grparr) {
 
     else if ($fldtype == 'lang') {
       $res = sqlStatement("SELECT * FROM lang_languages ORDER BY lang_description");
-      echo "  <select class='form-control' name='form_$i' id='form_$i'>\n";
+      echo "  <select class='form-control input-sm' name='form_$i' id='form_$i'>\n";
       while ($row = sqlFetchArray($res)) {
         echo "   <option value='" . attr($row['lang_description']) . "'";
         if ($row['lang_description'] == $fldvalue) echo " selected";
@@ -449,7 +549,7 @@ foreach ($GLOBALS_METADATA as $grpname => $grparr) {
         $order = "title, seq";
     }
       $res = sqlStatement("SELECT option_id, title FROM list_options WHERE list_id = ? AND activity=1 ORDER BY " . $order, array('apptstat'));
-      echo "  <select class='form-control' name='form_$i' id='form_$i'>\n";
+      echo "  <select class='form-control input-sm' name='form_$i' id='form_$i'>\n";
       if ($flddef ==" ") {
       $top_choice = "All";
       }else{
@@ -534,7 +634,7 @@ foreach ($GLOBALS_METADATA as $grpname => $grparr) {
 
     else if ($fldtype == 'all_code_types') {
       global $code_types;
-      echo "  <select class='form-control' name='form_$i' id='form_$i'>\n";
+      echo "  <select class='form-control input-sm' name='form_$i' id='form_$i'>\n";
       foreach (array_keys($code_types) as $code_key ) {
         echo "   <option value='" . attr($code_key) . "'";
         if ($code_key == $fldvalue) echo " selected";
@@ -547,7 +647,7 @@ foreach ($GLOBALS_METADATA as $grpname => $grparr) {
 
     else if ($fldtype == 'm_lang') {
       $res = sqlStatement("SELECT * FROM lang_languages  ORDER BY lang_description");
-      echo "  <select multiple name='form_{$i}[]' class='form-control' id='form_{$i}[]' size='3'>\n";
+      echo "  <select multiple name='form_{$i}[]' class='form-control input-sm' id='form_{$i}[]' size='3'>\n";
       while ($row = sqlFetchArray($res)) {
         echo "   <option value='" . attr($row['lang_description']) . "'";
         foreach ($glarr as $glrow) {
@@ -578,7 +678,7 @@ foreach ($GLOBALS_METADATA as $grpname => $grparr) {
       $themedir = "$webserver_root/interface/themes";
       $dh = opendir($themedir);
       if ($dh) {
-        echo "  <select class='form-control' name='form_$i' id='form_$i'>\n";
+        echo "  <select class='form-control input-sm' name='form_$i' id='form_$i'>\n";
         while (false !== ($tfname = readdir($dh))) {
           // Only show files that contain style_ as options
           //  Skip style_blue.css since this is used for
@@ -609,7 +709,7 @@ foreach ($GLOBALS_METADATA as $grpname => $grparr) {
       $themedir = "$webserver_root/interface/themes";
       $dh = opendir($themedir);
       if ($dh) {
-        echo "  <select class='form-control' name='form_$i' id='form_$i'>\n";
+        echo "  <select class='form-control input-sm' name='form_$i' id='form_$i'>\n";
         while (false !== ($tfname = readdir($dh))) {
           // Only show files that contain tabs_style_ as options
           if (!preg_match("/^tabs_style_.*\.css$/", $tfname)) continue;
@@ -632,7 +732,7 @@ foreach ($GLOBALS_METADATA as $grpname => $grparr) {
       if ($_GET['mode'] == "user") {
         $globalTitle = $globalValue;
       }
-      echo "  <select class='form-control' name='form_$i' id='form_$i'>\n";
+      echo "  <select class='form-control input-sm' name='form_$i' id='form_$i'>\n";
       for ($h = 0; $h < 24; ++$h) {
         echo "<option value='$h'";
         if ($h == $fldvalue) echo " selected";
@@ -644,6 +744,12 @@ foreach ($GLOBALS_METADATA as $grpname => $grparr) {
         echo "</option>\n";
       }
       echo "  </select>\n";
+    }
+    elseif ($fldtype == 'color') {
+      if ($_GET['mode'] == "user") {
+        $globalTitle = $globalValue;
+      }
+      echo "  <input type='color' class='form-control input-sm $fldid' name='form_$i' id='form_$i' value='" . attr($fldvalue) . "' />\n";
     }
     if ($_GET['mode'] == "user") {
       echo " </td>\n";
@@ -681,6 +787,31 @@ foreach ($GLOBALS_METADATA as $grpname => $grparr) {
 $(document).ready(function(){
   tabbify();
   enable_modals();
+
+//jquery for live theme selection, so once color picker picked up color css attributes will change.
+
+var primary_attributes = 'body_title, .body_top, .body_nav, .body_filler, .body_login, .table_bg, .bgcolor2, .textcolor1, .highlightcolor, .logobar';
+var secondary_attributes = "td, tr, .table, .bgcolor1,  ul.tabNav, .navbar, .nav, .dropdown, .navbar-header, input[type='submit'], ul.tabNav a";
+
+$('.primary_color').on("change", function () {
+var primary_color = $('.primary_color').val();
+$(primary_attributes).css('background-color', primary_color);
+});
+
+$('.primary_font_color').on("change", function () {
+var primary_font_color = $('.primary_font_color').val();
+$(primary_attributes).css('color', primary_font_color);
+});
+
+$('.secondary_color').on("change", function () {
+var secondary_color = $('.secondary_color').val();
+$(secondary_attributes).css('background-color', secondary_color);
+});
+
+$('.secondary_font_color').on("change", function () {
+var secondary_font_color = $('.secondary_font_color').val();
+$(secondary_attributes).css('color', secondary_font_color);
+});
 
   <?php // mdsupport - Highlight search results ?>
   $('.srch td').wrapInner("<mark></mark>");
